Moved Cardano.Ledger.Shelley.Orphans to Cardano.Ledger.Orphans Moved Cardano.Ledger.Shelley.PoolParams to Cardano.Ledger.PoolParams Moved Cardano.Ledger.Shelley.EpochBoundary to Cardano.Ledger.EpochBoundary Added deprecated versions in the old places that cause warnings, that tell what to import instead.
