+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/RewardProvenance.hs
@@ -24,8 +24,8 @@ import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Credential (Credential (..))
 import qualified Cardano.Ledger.Crypto as CC
 import Cardano.Ledger.Keys (KeyHash (..), KeyRole (..))
+import Cardano.Ledger.Orphans ()
 import Cardano.Ledger.SafeHash (SafeHash, unsafeMakeSafeHash)
-import Cardano.Ledger.Shelley.Orphans ()
 import Cardano.Ledger.Shelley.TxBody (PoolParams (..), RewardAcnt (..))
 import Control.DeepSeq (NFData)
 import Data.Aeson (FromJSON, ToJSON)
